requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-telegram-bot>=21.0
aiosqlite>=0.20.0
pydantic>=2.0
